pyproject.toml
[tool.poetry]
name = "lanchonete-autoatendimento"
version = "0.1.0"
description = "Sistema de autoatendimento para lanchonete"
authors = ["Seu Nome <seuemail@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
pillow = "^10.4.0"
drf-yasg = "^1.21.7"
setuptools = "^75.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
